window rendered

+++ src/player.rs
@@ -1,10 +1,10 @@
 pub struct Player {
-    pub size_w: u8,
-    pub size_h: u8,
-    pub pos_x: u8,
-    pub pos_y: u8,
-    pub sprite: [u8; 8],
-    pub score: i32
+    pub size_w: u32,
+    pub size_h: u32,
+    pub pos_x: u32,
+    pub pos_y: u32,
+    pub sprite: [u32; 8],
+    pub score: u32
 }
 
 impl Player {
@@ -41,10 +41,10 @@ impl Player {
     pub fn can_move(&mut self, bounds: sdl2::rect::Rect) -> bool {
         print!("x: {}\ty: {}\r", self.pos_x, self.pos_y );
        (
-           ( ( self.pos_x as i32 )  > bounds.x 
-           && ( self.pos_x as i32 + 8 ) < bounds.w )
-           || ( ( self.pos_y as i32 ) > bounds.y
-           && ( self.pos_y as i32 + 8 ) < bounds.h )
+           ( ( self.pos_x )  > bounds.x 
+           && ( self.pos_x + 8 ) < bounds.w )
+           || ( ( self.pos_y  ) > bounds.y
+           && ( self.pos_y + 8 ) < bounds.h )
        )
     }
